Refactor: - Created `Records` and `DataContainers` folders - Separated multi-class files into single files - Moved all classes that are not part of the nf-hello template or `CO2FootprintComputer` into subclasses - Adjusted imports
Test: Moved and renamed tests to fit new structure
