+++ plugins/nf-co2footprint/src/main/nextflow/co2footprint/DataContainers/DataMatrix.groovy
@@ -1,211 +1,9 @@
-package nextflow.co2footprint.utils
+package nextflow.co2footprint.DataContainers
 
 import groovy.util.logging.Slf4j
 
 import java.nio.file.Path
 import java.nio.file.Files
-
-/**
- * Bidirectional Map with maintained K-V pairs in both directions.
- *
- * Allows fast lookup of value by key and key by value.
- * Maintains insertion order and supports unique keys and values.
- *
- * @param <K> Key type
- * @param <V> Value type
- *
- * @author Josua Carl <josua.carl@uni-tuebingen.de>
- */
-class BiMap<K, V> {
-
-    // Map from key to value
-    private final Map<K, V> keyToValueMap = new LinkedHashMap<>()
-    // Map from value to key
-    private final Map<V, K> valueToKeyMap = new LinkedHashMap<>()
-
-    /**
-     * Construct a BiMap from a map or from parallel lists of keys and values.
-     * If both keys and values are provided, they must be unique and of equal length.
-     */
-    BiMap(Map<K,V> map = [:], List<K> keys = [], List<V> values = []) {
-        if (keys && values) {
-            assert  keys.size() == values.size()
-            assert  keys.unique() && values.unique()
-            for (int i = 0; i < keys.size(); i++) { this.put(keys[i], values[i]) }
-        }
-        else {
-            map.each { key, value -> this.put(key, value) }
-        }
-    }
-
-    /**
-     * Checks equality of two BiMap instances by comparing both key-value and value-key maps.
-     */
-    @Override
-    boolean equals(Object other) {
-        if ( other == null ) { return false }
-        if ( !(other instanceof BiMap) ) { return false }
-        BiMap bimap = (BiMap) other
-        if (this.keyToValueMap == bimap.keyToValueMap && this.valueToKeyMap == bimap.valueToKeyMap) { return true }
-        return false
-    }
-
-    /**
-     * Put a key-value pair into the bidirectional map.
-     * Both key and value must be unique.
-     */
-    void put(K key, V value)
-    {
-        keyToValueMap.put(key, value)
-        valueToKeyMap.put(value, key)
-    }
-
-    /**
-     * Get a value based on the key.
-     */
-    V getValue(K key) {
-        return keyToValueMap.get(key)
-    }
-
-    /**
-     * Get a key based on the value.
-     */
-    K getKey(V value) {
-        return valueToKeyMap.get(value)
-    }
-
-    /**
-     * Check if a key exists in the map.
-     */
-    boolean containsKey(K key) {
-        return keyToValueMap.containsKey(key)
-    }
-
-    /**
-     * Check if a value exists in the map.
-     */
-    boolean containsValue(V value) {
-        return valueToKeyMap.containsKey(value)
-    }
-
-    /**
-     * Filter keys using a closure and return a list of matching keys.
-     */
-    List filterKeys(Closure filterFunction) {
-        return keyToValueMap.keySet().stream().filter(filterFunction).toList()
-    }
-
-    /**
-     * Filter values using a closure and return a list of matching values.
-     */
-    List filterValues(Closure filterFunction) {
-        return valueToKeyMap.keySet().stream().filter(filterFunction).toList()
-    }
-
-    /**
-     * Remove a key-value pair based on the key.
-     * Returns the removed value.
-     */
-    V removeByKey(K key) {
-        final V value = keyToValueMap.remove(key)
-        valueToKeyMap.remove(value)
-        return value
-    }
-
-    /**
-     * Remove a key-value pair based on the value.
-     * Returns the removed key.
-     */
-    K removeByValue(V value) {
-        final K key = valueToKeyMap.remove(value)
-        keyToValueMap.remove(key)
-        return key
-    }
-
-    /**
-     * Remove all key-value pairs from the bidirectional map.
-     * Returns this BiMap for chaining.
-     */
-    BiMap<K,V> clear() {
-        keyToValueMap.clear()
-        valueToKeyMap.clear()
-        return this
-    }
-
-    /**
-     * Get a set of all keys in the bidirectional map.
-     */
-    Set<K> keySet() {
-        return keyToValueMap.keySet()
-    }
-
-    /**
-     * Get a set of all values in the bidirectional map.
-     */
-    Set<V> valueSet() {
-        return valueToKeyMap.keySet()
-    }
-
-    /**
-    * Get the number of key-value pairs in the map.
-    */
-    Integer size() {
-        return keyToValueMap.size()
-    }
-
-    /**
-     * Return a new BiMap sorted by values.
-     */
-    BiMap<K,V> sortByValues() {
-        final List<V> values = valueToKeyMap.keySet().sort()
-        final List<K> keys = values.collect { value -> valueToKeyMap[value]}
-
-        return new BiMap(null, keys, values)
-    }
-
-    /**
-     * Return a new BiMap sorted by keys.
-     */
-    BiMap<K,V> sortByKeys() {
-        final List<K> keys = keyToValueMap.keySet().sort()
-        final List<V> values = keys.collect { key -> keyToValueMap[key]}
-
-        return new BiMap(null, keys, values)
-    }
-
-    /**
-     * String representation of the BiMap (shows key-to-value mapping).
-     */
-    String toString() {
-        return keyToValueMap.toString()
-    }
-}
-
-
-
-/**
- * Interface for a Table/Matrix.
- * Defines the basic structure and required methods for a matrix-like data container.
- *
- * @author Josua Carl <josua.carl@uni-tuebingen.de>
- */
-interface Matrix {
-    // The matrix data as a list of lists (rows of columns)
-    List<List<Object>> data = []
-    // Column index mapping (column name -> index)
-    BiMap<Object, Integer> columnIndex = [:] as BiMap
-    // Row index mapping (row name -> index)
-    BiMap<Object, Integer> rowIndex = [:] as BiMap
-
-    // Get a value from the matrix at the specified row and column.
-    Object get(Object row, Object column, boolean rowRawIndex, boolean columnRawIndex)
-
-    // Select a submatrix by rows and columns.
-    Matrix select(LinkedHashSet<Object> rows, LinkedHashSet<Object> columns)
-
-    // Set a value in the matrix at the specified row and column.
-    void set(Object value, Object row, Object column, boolean rowRawIndex, boolean columnRawIndex)
-}
 
 /**
  * DataMatrix / Table Base Class.
