+++ plugins/nf-co2footprint/src/main/nextflow/co2footprint/FileCreators/CO2FootprintSummary.groovy
@@ -2,7 +2,7 @@ package nextflow.co2footprint.FileCreators
 
 import groovy.util.logging.Slf4j
 import groovyx.gpars.agent.Agent
-import nextflow.co2footprint.CO2EquivalencesRecord
+import nextflow.co2footprint.Records.CO2EquivalencesRecord
 import nextflow.co2footprint.CO2FootprintConfig
 import nextflow.co2footprint.utils.Converter
 import nextflow.trace.TraceHelper
