+++ plugins/nf-co2footprint/src/main/nextflow/co2footprint/Records/CO2RecordAggregator.groovy
@@ -1,4 +1,4 @@
-package nextflow.co2footprint
+package nextflow.co2footprint.Records
 
 import groovy.transform.CompileStatic
 import groovy.util.logging.Slf4j
