+++ plugins/nf-co2footprint/src/test/nextflow/co2footprint/CO2FootprintObserverTest.groovy
@@ -2,6 +2,7 @@ package nextflow.co2footprint
 
 import nextflow.NextflowMeta
 import nextflow.Session
+import nextflow.co2footprint.Records.CO2EquivalencesRecord
 import nextflow.executor.NopeExecutor
 import nextflow.processor.TaskHandler
 import nextflow.processor.TaskId
